Add files via upload
This was a large-scale Netflix movie data analysis project, and I wanted to understand how the data worked. 
The data was massive and riddled with problems… everything was piled on top of each other: 
user data, movies, and sessions all in one sheet, so I had to manually separate them. 
I used Power Query and split the data into three sheets: 
one for users, one for movies, and one for sessions, then I properly linked them.

I changed the incorrect data types and organized everything until the data was clean and ready for analysis. 
After that, I started to see the complete picture: the number of movies I had, the number of users,
the number of sessions, viewing activity over the months, and the total revenue increasing each month.

I also analyzed each user's subscription type, the video quality they watched,
the movie genres (action, drama, comedy, etc.), the producing countries, the countries the user watched from, 
and whether the movie was family-friendly. I even found out where the user watched from… mobile? No laptop? No TV?

In short… this was a challenging project, but it made a huge difference in my understanding of data analysis, 
cleaning, and linking, and it yielded powerful insights.